+++ packages/database/src/operations/index.ts
@@ -2,4 +2,4 @@ export * from './users';
 
 export * from './sync';
 
-export * from './insert';
+export * from './dmhy';
